take different pointer sizes into account
The casting that was used here resulted in compiler warnings. Let's use
intptr_t of architecture dependent size.

+++ plugins/dejitter.c
@@ -25,6 +25,7 @@
 #include <stdlib.h>
 #include <string.h>
 #include <limits.h>
+#include <stdint.h>
 
 #include <stdio.h>
 
@@ -179,7 +180,7 @@ static int dejitter_limit(struct tslib_module_info *inf, char *str, void *data)
 		return -1;
 
 	errno = err;
-	switch ((int)data) {
+	switch ((int)(intptr_t)data) {
 	case 1:
 		djt->delta = v;
 		break;
